# docs/reference/openapi-compliance.yaml
openapi: 3.1.0
info:
  title: Luqen Compliance Service
  description: >-
    Accessibility compliance rule engine — check WCAG issues against 60+
    country-specific legal requirements. Provides CRUD for jurisdictions,
    regulations, and requirements, a compliance checker, update proposal
    workflow, monitored sources, webhooks, user/client management, and OAuth
    token issuance.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/jurisdictions:
    get:
      operationId: listJurisdictions
      tags: [jurisdictions]
      summary: List jurisdictions
      description: Returns a paginated list of jurisdictions, optionally filtered.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: type
          in: query
          schema:
            type: string
            enum: [supranational, country, state]
        - name: parentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated jurisdiction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedJurisdictions"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createJurisdiction
      tags: [jurisdictions]
      summary: Create a jurisdiction
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJurisdictionInput"
      responses:
        "201":
          description: Jurisdiction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jurisdiction"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/regulations:
    get:
      operationId: listRegulations
      tags: [regulations]
      summary: List regulations
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: jurisdictionId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, draft, repealed]
        - name: scope
          in: query
          schema:
            type: string
            enum: [public, private, all]
      responses:
        "200":
          description: Paginated regulation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRegulations"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createRegulation
      tags: [regulations]
      summary: Create a regulation
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegulationInput"
      responses:
        "201":
          description: Regulation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List monitored sources
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonitoredSource"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createSource
      tags: [sources]
      summary: Create a monitored source
      description: Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceInput"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredSource"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/sources/{id}:
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete a monitored source
      description: Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "204":
          description: Source deleted
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sources/scan:
    post:
      operationId: scanSources
      tags: [sources]
      summary: Trigger a scan of all monitored sources
      description: >-
        Fetches each monitored source, computes a SHA-256 content hash, and
        creates update proposals for any sources whose content has changed.
        Requires `admin` scope.
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSourcesResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ────────────────────────────────────────────────────────────────────────────
  # Webhooks
  # ────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
    XOrgId:
      name: X-Org-Id
      in: header
      description: >-
        Optional organization context. When set, scopes data operations to
        this organization.
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return (default 50).
      schema:
        type: integer
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of items to skip (default 0).
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, statusCode]
      properties:
        error:
          type: string
        statusCode:
          type: integer

    # ── Health ─────────────────────────────────────────────────────────────────
...
    Jurisdiction:
      type: object
      required: [id, name, type, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [supranational, country, state]
        parentId:
          type: string
        iso3166:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateJurisdictionInput:
      type: object
      required: [name, type]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [supranational, country, state]
        parentId:
          type: string
        iso3166:
          type: string
...
    PaginatedJurisdictions:
      type: object
      required: [data, total, limit, offset]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Jurisdiction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ── Regulation ─────────────────────────────────────────────────────────────
    Regulation:
      type: object
      required:
        - id
        - jurisdictionId
        - name
        - shortName
        - reference
        - url
        - enforcementDate
        - status
        - scope
        - sectors
        - description
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        jurisdictionId:
          type: string
        name:
          type: string
        shortName:
          type: string
        reference:
          type: string
        url:
          type: string
          format: uri
        enforcementDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, draft, repealed]
        scope:
          type: string
          enum: [public, private, all]
        sectors:
          type: array
          items:
            type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRegulationInput:
      type: object
      required:
        - jurisdictionId
        - name
        - shortName
        - reference
        - url
        - enforcementDate
        - status
        - scope
        - sectors
        - description
      properties:
        id:
          type: string
        jurisdictionId:
          type: string
        name:
          type: string
        shortName:
          type: string
        reference:
          type: string
        url:
          type: string
          format: uri
        enforcementDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, draft, repealed]
        scope:
          type: string
          enum: [public, private, all]
        sectors:
          type: array
          items:
            type: string
        description:
          type: string
...
    PaginatedRegulations:
      type: object
      required: [data, total, limit, offset]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Regulation"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ── Requirement ────────────────────────────────────────────────────────────
...
    UpdateProposal:
      type: object
      required: [id, source, detectedAt, type, summary, proposedChanges, status, createdAt]
      properties:
        id:
          type: string
        source:
          type: string
        detectedAt:
          type: string
          format: date-time
        type:
          type: string
          enum: [new_regulation, amendment, repeal, new_requirement, new_jurisdiction]
        affectedRegulationId:
          type: string
        affectedJurisdictionId:
          type: string
        summary:
          type: string
        proposedChanges:
          $ref: "#/components/schemas/ProposedChange"
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedBy:
          type: string
        reviewedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ProposedChange:
      type: object
      required: [action, entityType]
      properties:
        action:
          type: string
          enum: [create, update, delete]
        entityType:
          type: string
          enum: [jurisdiction, regulation, requirement]
        entityId:
          type: string
        before:
          type: object
          additionalProperties: true
        after:
          type: object
          additionalProperties: true
...
    MonitoredSource:
      type: object
      required: [id, name, url, type, schedule, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [html, rss, api]
        schedule:
          type: string
          enum: [daily, weekly, monthly]
        lastCheckedAt:
          type: string
          format: date-time
        lastContentHash:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateSourceInput:
      type: object
      required: [name, url, type, schedule]
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [html, rss, api]
        schedule:
          type: string
          enum: [daily, weekly, monthly]
    ScanSourcesResponse:
      type: object
      required: [scanned, proposalsCreated, proposals]
      properties:
        scanned:
          type: integer
          description: Number of sources scanned.
        proposalsCreated:
          type: integer
          description: Number of change-detected proposals created.
        proposals:
          type: array
          items:
            $ref: "#/components/schemas/UpdateProposal"

    # ── Webhook ────────────────────────────────────────────────────────────────
...